OO-7069: add media part owner and workaround for missing left joins

+++ src/main/java/org/olat/modules/cemedia/ui/ConfirmDeleteMediaController.java
@@ -77,9 +77,19 @@ public class ConfirmDeleteMediaController extends FormBasicController {
 
 	@Override
 	protected void formOK(UserRequest ureq) {
+		boolean stillUsed = false;
 		for(MediaRow rowToDelete:rowsToDelete) {
 			Media media = mediaService.getMediaByKey(rowToDelete.getKey());
-			mediaService.deleteMedia(media);
+			if(mediaService.isUsed(media)) {
+				stillUsed = true;
+			} else {
+				mediaService.deleteMedia(media);
+			
+			}
+		}
+		
+		if(stillUsed) {
+			showWarning("warning.still.used");
 		}
 		fireEvent(ureq, Event.DONE_EVENT);
 	}
